[Backend] Refactor the proxy
The proxy got some well-deserved code splitting
