+++ src/canvasapis/users.go
@@ -1,7 +1,7 @@
 package canvasapis
 
 import (
-	"github.com/iamtheyammer/canvas-grade-calculator/backend/src/canvasapis/services"
+	"github.com/iamtheyammer/canvas-grade-calculator/backend/src/canvasapis/services/users"
 	"github.com/iamtheyammer/canvas-grade-calculator/backend/src/util"
 	"github.com/julienschmidt/httprouter"
 	"net/http"
@@ -14,7 +14,7 @@ func GetOwnUserProfileHandler(w http.ResponseWriter, r *http.Request, _ httprout
 		return
 	}
 
-	resp, body, err := services.GetOwnUserProfile(rd)
+	resp, body, err := users.GetSelfProfile(rd)
 	if err != nil {
 		util.SendInternalServerError(w)
 	}
